상수에 대한 어셈블리의존? 추가 및 기존 상수 사용중인 코드에 using 추가
