+++ Assets/_Dev/Scripts/Util/define.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e1fcbc6cfa4d5d24295425ea81781d05
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
