+++ Assets/_Dev/Scripts/Util/define/WheelMaker.Define.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "WheelMaker.Define"
+}
